Move type component checker to FindPossibleSecretsVisitor
